LayoutTests/imported/w3c/web-platform-tests/streams/piping/close-propagation-forward.any-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream

Harness Error (FAIL), message = Unhandled rejection: cancel() called on a reader owned by no readable stream

FAIL Closing must be propagated forward: starts closed; preventClose omitted; fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: starts closed; preventClose omitted; rejected close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: starts closed; preventClose = undefined (falsy); fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: starts closed; preventClose = null (falsy); fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: starts closed; preventClose = false (falsy); fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: starts closed; preventClose = 0 (falsy); fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: starts closed; preventClose = -0 (falsy); fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: starts closed; preventClose = NaN (falsy); fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: starts closed; preventClose =  (falsy); fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
PASS Closing must be propagated forward: starts closed; preventClose = true (truthy) 
PASS Closing must be propagated forward: starts closed; preventClose = a (truthy) 
PASS Closing must be propagated forward: starts closed; preventClose = 1 (truthy) 
PASS Closing must be propagated forward: starts closed; preventClose = Symbol() (truthy) 
PASS Closing must be propagated forward: starts closed; preventClose = [object Object] (truthy) 
PASS Closing must be propagated forward: starts closed; preventClose = true, preventAbort = true 
PASS Closing must be propagated forward: starts closed; preventClose = true, preventAbort = true, preventCancel = true 
FAIL Closing must be propagated forward: becomes closed asynchronously; preventClose omitted; fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: becomes closed asynchronously; preventClose omitted; rejected close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
PASS Closing must be propagated forward: becomes closed asynchronously; preventClose = true 
FAIL Closing must be propagated forward: becomes closed asynchronously; dest never desires chunks; preventClose omitted; fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: becomes closed asynchronously; dest never desires chunks; preventClose omitted; rejected close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
PASS Closing must be propagated forward: becomes closed asynchronously; dest never desires chunks; preventClose = true 
FAIL Closing must be propagated forward: becomes closed after one chunk; preventClose omitted; fulfilled close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
FAIL Closing must be propagated forward: becomes closed after one chunk; preventClose omitted; rejected close promise promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
PASS Closing must be propagated forward: becomes closed after one chunk; preventClose = true 
PASS Closing must be propagated forward: shutdown must not occur until the final write completes 
PASS Closing must be propagated forward: shutdown must not occur until the final write completes; preventClose = true 
PASS Closing must be propagated forward: shutdown must not occur until the final write completes; becomes closed after first write 
PASS Closing must be propagated forward: shutdown must not occur until the final write completes; becomes closed after first write; preventClose = true 
FAIL Closing must be propagated forward: erroring the writable while flushing pending writes should error pipeTo promise_test: Unhandled rejection with value: object "TypeError: ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)"
